+++ test/test_config_files.rb
@@ -48,13 +48,6 @@ class TestConfigFiles < KwardTestCase
     end
   end
 
-  def test_banner_enabled_defaults_to_true_and_only_false_disables_it
-    assert_equal true, Kward::ConfigFiles.banner_enabled?({})
-    assert_equal true, Kward::ConfigFiles.banner_enabled?("banner" => {})
-    assert_equal true, Kward::ConfigFiles.banner_enabled?("banner" => { "enabled" => true })
-    assert_equal false, Kward::ConfigFiles.banner_enabled?("banner" => { "enabled" => false })
-  end
-
   def test_workspace_guardrails_enabled_defaults_to_true_and_only_false_disables_it
     assert_equal true, Kward::ConfigFiles.workspace_guardrails_enabled?({})
     assert_equal true, Kward::ConfigFiles.workspace_guardrails_enabled?("tools" => {})